Former-commit-id: 2d7e09d5c45729a2e823624330b11c1746501661
Former-commit-id: 8945323f620ccbaba0b587f7ab4c05bb8e32c0ab

+++ automationtests/src/main/java/io/mosip/authentication/fw/util/AuthTestsUtil.java
@@ -31,19 +31,16 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.testng.Reporter;
 import org.testng.annotations.AfterClass;
-import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeClass;
-import org.testng.annotations.BeforeSuite;
-
 import com.google.common.base.Verify;
 import com.ibm.icu.text.Transliterator;
 
 import io.mosip.authentication.fw.dto.OutputValidationDto;
 import io.mosip.authentication.fw.precon.JsonPrecondtion;
+import io.mosip.authentication.fw.precon.MessagePrecondtion;
 import io.mosip.authentication.fw.precon.XmlPrecondtion;
 import io.mosip.authentication.idRepository.fw.util.IdRepoTestsUtil;
 import io.mosip.service.BaseTestCase;
-import io.mosip.testrunner.MosipTestRunner;
 import io.restassured.response.Response;
  
 /**
@@ -63,6 +60,7 @@ public class AuthTestsUtil extends BaseTestCase {
 	protected static String responseJsonToVerifyDigtalSignature;
 	protected static String responseDigitalSignatureValue;
 	protected static String responseDigitalSignatureKey="response-signature";
+	protected static String logFileName="id-auth.log";
 	
 	/**
 	 * The method will get current test execution folder
@@ -176,14 +174,16 @@ public class AuthTestsUtil extends BaseTestCase {
 				if (listOfFiles[j].getName().contains(keywordToFind)) {
 					FileOutputStream fos = new FileOutputStream(
 							listOfFiles[j].getParentFile() + "/" + generateOutputFileKeyword + ".json");
-					String responseJson = "";
+					Response response;
 					if (code == 0)
-						responseJson = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath,cookieName,cookieValue);
+						response = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath,cookieName,cookieValue);
 					else
-						responseJson = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath, code,cookieName,cookieValue);
+						response = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath, code,cookieName,cookieValue);
+					responseJsonToVerifyDigtalSignature=response.asString();
+					responseDigitalSignatureValue=response.getHeader(responseDigitalSignatureKey);
 					Reporter.log("<b><u>Actual Response Content: </u></b>(EndPointUrl: " + urlPath + ") <pre>"
-							+ ReportUtil.getTextAreaJsonMsgHtml(responseJson) + "</pre>");
-					fos.write(responseJson.getBytes());
+							+ ReportUtil.getTextAreaJsonMsgHtml(response.asString()) + "</pre>");
+					fos.write(response.asString().getBytes());
 					fos.flush();
 					fos.close();
 				}
@@ -212,17 +212,17 @@ public class AuthTestsUtil extends BaseTestCase {
 				if (listOfFiles[j].getName().contains(keywordToFind)) {
 					FileOutputStream fos = new FileOutputStream(
 							listOfFiles[j].getParentFile() + "/" + generateOutputFileKeyword + ".json");
-					String responseJson = "";
+					Response responseJson;
 					if (code == 0)
 						responseJson = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath,cookieName,cookieValue);
 					else
 						responseJson = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath, code,cookieName,cookieValue);
 					Reporter.log("<b><u>Actual Response Content: </u></b>(EndPointUrl: " + urlPath + ") <pre>"
-							+ ReportUtil.getTextAreaJsonMsgHtml(responseJson) + "</pre>");
-					fos.write(responseJson.getBytes());
+							+ ReportUtil.getTextAreaJsonMsgHtml(responseJson.asString()) + "</pre>");
+					fos.write(responseJson.asString().getBytes());
 					fos.flush();
 					fos.close();
-					return responseJson.toString();
+					return responseJson.asString().toString();
 				}
 			}
 			return "NoResponse";
@@ -232,7 +232,7 @@ public class AuthTestsUtil extends BaseTestCase {
 		}
 	}
 	
-	protected String postRequestWithCookie(String filename, String url, int expCode,String cookieName,String cookieValue) {
+	protected Response postRequestWithCookie(String filename, String url, int expCode,String cookieName,String cookieValue) {
 		Response response=null;
 		try {
 			JSONObject objectData = (JSONObject) new JSONParser().parse(new FileReader(filename));
@@ -244,10 +244,10 @@ public class AuthTestsUtil extends BaseTestCase {
 			objMap.put("Status Code", objList);
 			Reporter.log(ReportUtil.getOutputValiReport(objMap));
 			Verify.verify(OutputValidationUtil.publishOutputResult(objMap));
-			return response.asString();
+			return response;
 		} catch (Exception e) {
 			IDASCRIPT_LOGGER.error("Exception: " + e);
-			return response.asString();
+			return response;
 		}
 	}
 	
@@ -268,19 +268,19 @@ public class AuthTestsUtil extends BaseTestCase {
 				if (listOfFiles[j].getName().contains(keywordToFind)) {
 					FileOutputStream fos = new FileOutputStream(
 							listOfFiles[j].getParentFile() + "/" + generateOutputFileKeyword + ".json");
-					String responseJson = "";
+					Response responseJson;
 					if (code == 0)
 						responseJson = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath,cookieName,cookieValue);
 					else
 						responseJson = postRequestWithCookie(listOfFiles[j].getAbsolutePath(), urlPath, code,cookieName,cookieValue);
-					if (responseJson.contains("Invalid UIN")) {
+					if (responseJson.asString().contains("Invalid UIN")) {
 						fos.flush();
 						fos.close();
 						return false;
 					} else {
 						Reporter.log("<b><u>Actual Response Content: </u></b>(EndPointUrl: " + urlPath + ") <pre>"
-								+ ReportUtil.getTextAreaJsonMsgHtml(responseJson) + "</pre>");
-						fos.write(responseJson.getBytes());
+								+ ReportUtil.getTextAreaJsonMsgHtml(responseJson.asString()) + "</pre>");
+						fos.write(responseJson.asString().getBytes());
 						fos.flush();
 						fos.close();
 						return true;
@@ -626,8 +626,11 @@ public class AuthTestsUtil extends BaseTestCase {
 		try {
 			for (int j = 0; j < listOfFiles.length; j++) {
 				if (listOfFiles[j].getName().contains(keywordinFile))
-					JsonPrecondtion.parseAndwriteJsonFile(listOfFiles[j].getAbsolutePath(), fieldvalue,
+				{
+					MessagePrecondtion jsonPrecon = new JsonPrecondtion();
+					jsonPrecon.parseAndWriteFile(listOfFiles[j].getAbsolutePath(), fieldvalue,
 							listOfFiles[j].getAbsolutePath(), propFileName);
+				}
 			}
 			return true;
 		} catch (Exception e) {
@@ -808,7 +811,7 @@ public class AuthTestsUtil extends BaseTestCase {
 	 * @param filePath
 	 * @param map
 	 */
-	public void updateMappingDic(String filePath, Map<String, String> map) {
+	public static void updateMappingDic(String filePath, Map<String, String> map) {
 		try {
 			FileInputStream in = new FileInputStream(filePath);
 			Properties props = new Properties();
@@ -939,8 +942,14 @@ public class AuthTestsUtil extends BaseTestCase {
 	 */
 	@AfterClass
 	public static void authTestTearDown() {
-		IDASCRIPT_LOGGER.info("Terminating authpartner demo application...");
-		AuthPartnerProcessor.authPartherProcessor.destroyForcibly();
+		if (AuthenticationTestsListener.currentTestMethod.toString().contains(".authentication.tests")
+				|| AuthenticationTestsListener.currentTestMethod.toString()
+						.contains(".authentication.idRepository.tests")
+				|| AuthenticationTestsListener.currentTestMethod.toString()
+						.contains(".authentication.idRepository.prerequiste")) {
+			IDASCRIPT_LOGGER.info("Terminating authpartner demo application...");
+			AuthPartnerProcessor.authPartherProcessor.destroyForcibly();
+		}
 	}
 	
 	/**
@@ -948,13 +957,22 @@ public class AuthTestsUtil extends BaseTestCase {
 	 */
 	@BeforeClass
 	public static void wakeDemoApp() {
-		IDASCRIPT_LOGGER.info("Starting authpartner demo application...");
-		AuthPartnerProcessor.startProcess();
-		/*createBatOrShFileForDemoApp();
-		if (getOSType().toString().equals("WINDOWS"))
-			batDemoAppRunner();
-		else if (getOSType().toString().equals("OTHERS"))
-			shDemoAppRunner();*/
+		if (AuthenticationTestsListener.currentTestMethod.toString().contains(".authentication.tests")
+				|| AuthenticationTestsListener.currentTestMethod.toString()
+						.contains(".authentication.idRepository.tests")
+				|| AuthenticationTestsListener.currentTestMethod.toString()
+						.contains(".authentication.idRepository.prerequiste")) {
+			IDASCRIPT_LOGGER.info("Starting authpartner demo application...");
+			AuthPartnerProcessor.startProcess();
+		}
+	}
+	
+	public static void regenerateVID() {
+		if (!AuthenticationTestsListener.currentTestMethod.toString()
+				.contains("authentication.idRepository.prerequiste")) {
+			IDASCRIPT_LOGGER.info("Regenerating Used VID...");
+			VIDUtil.regenerateVID();
+		}
 	}
 	
 	public static void initiateAuthTest() {
@@ -1188,15 +1206,14 @@ public class AuthTestsUtil extends BaseTestCase {
 				MediaType.APPLICATION_JSON,cookieName);
 	}
 	
-	protected String postRequestWithCookie(String filename, String url,String cookieName, String cookieValue) {
+	protected Response postRequestWithCookie(String filename, String url,String cookieName, String cookieValue) {
 		try {
 			JSONObject objectData = (JSONObject) new JSONParser().parse(new FileReader(filename));
 			return RestClient
-					.postRequestWithCookie(url, objectData.toJSONString(), MediaType.APPLICATION_JSON, MediaType.APPLICATION_JSON,cookieName,cookieValue)
-					.asString();
+					.postRequestWithCookie(url, objectData.toJSONString(), MediaType.APPLICATION_JSON, MediaType.APPLICATION_JSON,cookieName,cookieValue);
 		} catch (Exception e) {
 			IDASCRIPT_LOGGER.error("Exception: " + e);
-			return e.toString();
+			return null;
 		}
 	}
 	protected String postRequestWithCookie(String filename, String url,String cookieName, String cookieValue,int code) {
@@ -1210,7 +1227,7 @@ public class AuthTestsUtil extends BaseTestCase {
 			return e.toString();
 		}
 	}
-	protected String postStrContentRequestWithCookie(String content, String url,String cookieName, String cookieValue) {
+	protected static String postStrContentRequestWithCookie(String content, String url,String cookieName, String cookieValue) {
 		try {
 			return RestClient
 					.postRequestWithCookie(url, content, MediaType.APPLICATION_JSON, MediaType.APPLICATION_JSON,cookieName,cookieValue)
@@ -1269,7 +1286,17 @@ public class AuthTestsUtil extends BaseTestCase {
 	public static String getVidRequestContent() {
 		try {
 			return getContentFromFile(new File(RunConfigUtil.getResourcePath()
-					+ "ida/VIDData/VIDGeneration/VIDGenerate/vid-request.json"));
+					+ "ida/VIDData/VIDGeneration/VIDGenerate_smoke/vid-request.json"));
+		} catch (Exception e) {
+			IDASCRIPT_LOGGER.error("Exception Occured in getting the VID request file" + e.getMessage());
+			return e.getMessage();
+		}
+	}
+	
+	public static String getVidRequestContentTemplate() {
+		try {
+			return getContentFromFile(new File(RunConfigUtil.getResourcePath()
+					+ "ida/TestData/VIDData/VIDGeneration/input/vid-request.json"));
 		} catch (Exception e) {
 			IDASCRIPT_LOGGER.error("Exception Occured in getting the VID request file" + e.getMessage());
 			return e.getMessage();
@@ -1286,7 +1313,7 @@ public class AuthTestsUtil extends BaseTestCase {
 	 * @param code
 	 * @return true or false
 	 */
-	protected String postRequestAndGetResponseForVIDGeneration(String content, String urlPath, String cookieName,
+	public static String postRequestAndGetResponseForVIDGeneration(String content, String urlPath, String cookieName,
 			String cookieValue) {
 		try {
 			return postStrContentRequestWithCookie(content, urlPath, cookieName, cookieValue);
@@ -1310,7 +1337,7 @@ public class AuthTestsUtil extends BaseTestCase {
 	public static void copyAuthTestResource() {
 		try {
 			File source = new File(RunConfigUtil.getGlobalResourcePath() + "/ida");
-			File destination = new File(RunConfigUtil.getGlobalResourcePath() + "/AuthenticationTestResource");
+			File destination = new File(RunConfigUtil.getGlobalResourcePath() + "/"+RunConfigUtil.resourceFolderName);
 			FileUtils.copyDirectoryToDirectory(source, destination);
 			IDASCRIPT_LOGGER.info("Copied the authentication test resource successfully");
 		} catch (Exception e) {
@@ -1319,12 +1346,12 @@ public class AuthTestsUtil extends BaseTestCase {
 	}
 
 	public static void removeOldAuthTestResource() {
-		File authTestFile = new File(RunConfigUtil.getGlobalResourcePath() + "/AuthenticationTestResource");
+		File authTestFile = new File(RunConfigUtil.getGlobalResourcePath() + "/"+RunConfigUtil.resourceFolderName);
 		if (authTestFile.exists())
 			if (FileUtil.deleteDirectory(authTestFile))
-				IDASCRIPT_LOGGER.info("Old AuthenticationTestResource folder successfully deleted!!");
+				IDASCRIPT_LOGGER.info("Old "+RunConfigUtil.resourceFolderName+" folder successfully deleted!!");
 			else
-				IDASCRIPT_LOGGER.error("Old AuthenticationTestResource folder not deleted.");
+				IDASCRIPT_LOGGER.error("Old "+RunConfigUtil.resourceFolderName+" folder not deleted.");
 	}
 	/**
 	 * The method will get request and generate output file with return repose
